# backend/docs/swagger.yaml
basePath: /
definitions:
  application.ClaimRequest:
    properties:
      checkpoint_val:
        type: integer
      player_id:
        type: string
    required:
    - checkpoint_val
    - player_id
    type: object
  application.ClaimResponse:
    properties:
      checkpoint_val:
        type: integer
      claimed_at:
        type: string
      id:
        type: string
      reward_name:
        type: string
    type: object
  application.EnterRequest:
    properties:
      nickname:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - nickname
    type: object
  application.EnterResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      is_new:
        description: Only for enter
        type: boolean
      nickname:
        type: string
      total_points:
        type: integer
    type: object
  application.GetHistoryResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/application.RewardHistoryDTO'
        type: array
    type: object
  application.GlobalHistoryResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/application.GlobalSpinLogDTO'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  application.GlobalSpinLogDTO:
    properties:
      created_at:
        type: string
      id:
        type: string
      player_nickname:
        type: string
      points_gained:
        type: integer
      source:
        type: string
    type: object
  application.PersonalHistoryResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/application.PersonalSpinLogDTO'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  application.PersonalSpinLogDTO:
    properties:
      created_at:
        type: string
      id:
        type: string
      points_gained:
        type: integer
      source:
        type: string
    type: object
  application.ProfileResponse:
    properties:
      claimed_checkpoints:
        description: Phase 6
        items:
          type: integer
        type: array
      created_at:
        type: string
      id:
        type: string
      nickname:
        type: string
      total_points:
        type: integer
    type: object
  application.RewardHistoryDTO:
    properties:
      checkpoint_val:
        type: integer
      claimed_at:
        type: string
      id:
        type: string
      reward_description:
        type: string
      reward_name:
        type: string
    type: object
  application.SpinErrorResponse:
    properties:
      code:
        example: DAILY_LIMIT_EXCEEDED
        type: string
      message:
        example: Daily spin limit reached
        type: string
      remaining_spins:
        example: 0
        type: integer
    type: object
  application.SpinRequest:
    properties:
      player_id:
        example: uuid-123
        type: string
    type: object
  application.SpinResponse:
    properties:
      points_gained:
        example: 500
        type: integer
      spin_id:
        example: uuid-456
        type: string
      total_points_after:
        example: 1500
        type: integer
    type: object
host: localhost:3001
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A spin game API built with Go and Fiber
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Spin Head API
  version: "1.0"
paths:
  /game/spin:
    post:
      consumes:
      - application/json
      description: Perform a spin for the player (max 10 spins/day)
      parameters:
      - description: Spin request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/application.SpinRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/application.SpinResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/application.SpinErrorResponse'
        "404":
          description: Player not found
          schema:
            $ref: '#/definitions/application.SpinErrorResponse'
        "429":
          description: Daily limit exceeded
          schema:
            $ref: '#/definitions/application.SpinErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/application.SpinErrorResponse'
      summary: Execute a spin
      tags:
      - Game
  /history/{player_id}:
    get:
      consumes:
      - application/json
      description: Get paginated spin history for a specific player
      parameters:
      - description: Player ID
        in: path
        name: player_id
        required: true
        type: string
      - default: 20
        description: Number of items per page
        in: query
        name: limit
        type: integer
      - default: 0
        description: Number of items to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/application.PersonalHistoryResponse'
        "400":
          description: Bad Request
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      summary: Get personal spin history
      tags:
      - History
  /history/global:
    get:
      consumes:
      - application/json
      description: Get paginated global spin history with player nicknames
      parameters:
      - default: 20
        description: Number of items per page
        in: query
        name: limit
        type: integer
      - default: 0
        description: Number of items to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/application.GlobalHistoryResponse'
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      summary: Get global spin history
      tags:
      - History
  /players/{id}:
    get:
      consumes:
      - application/json
      description: Get player profile information by player ID
      parameters:
      - description: Player ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Player profile
          schema:
            $ref: '#/definitions/application.ProfileResponse'
        "400":
          description: Bad request
          schema:
            type: object
        "404":
          description: Player not found
          schema:
            type: object
        "500":
          description: Internal server error
          schema:
            type: object
      summary: Get player profile
      tags:
      - Players
  /players/enter:
    post:
      consumes:
      - application/json
      description: Enter an existing player or create a new one with the given nickname
      parameters:
      - description: Player enter request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/application.EnterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Existing player found
          schema:
            $ref: '#/definitions/application.EnterResponse'
        "201":
          description: New player created
          schema:
            $ref: '#/definitions/application.EnterResponse'
        "400":
          description: Bad request
          schema:
            type: object
        "500":
          description: Internal server error
          schema:
            type: object
      summary: Enter or resume a player
      tags:
      - Players
  /rewards/{player_id}:
    get:
      consumes:
      - application/json
      description: Get all rewards claimed by a player
      parameters:
      - description: Player ID
        in: path
        name: player_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/application.GetHistoryResponse'
        "400":
          description: Invalid player ID
          schema:
            type: object
        "500":
          description: Internal server error
          schema:
            type: object
      summary: Get reward claim history
      tags:
      - Rewards
  /rewards/claim:
    post:
      consumes:
      - application/json
      description: Claim a reward when player reaches a checkpoint
      parameters:
      - description: Claim reward request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/application.ClaimRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/application.ClaimResponse'
        "400":
          description: Insufficient points or invalid checkpoint
          schema:
            type: object
        "404":
          description: Player not found
          schema:
            type: object
        "409":
          description: Already claimed
          schema:
            type: object
        "500":
          description: Internal server error
          schema:
            type: object
      summary: Claim reward at checkpoint
      tags:
      - Rewards
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"